# docs/patra_openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Patra Knowledge Graph API",
    "version": "1.0.0",
    "description": "API for interacting with the Patra Knowledge Graph, providing comprehensive model card management, datasheet operations, and model deployment tracking capabilities.",
    "contact": {
      "name": "Patra KG Team",
      "url": "https://github.com/patra-kg"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    }
  },
  "servers": [
    {
      "url": "http://localhost:5002",
      "description": "Development server"
    },
    {
      "url": "https://api.patra-kg.org",
      "description": "Production server"
    }
  ],
  "paths": {
...
    "/modelcard": {
      "post": {
        "summary": "Upload model card",
        "description": "Uploads a new model card to the Patra Knowledge Graph",
        "operationId": "post_upload_model_card",
        "tags": ["Model Cards"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ModelCard"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Model card successfully uploaded or already exists",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "type": "object",
                      "properties": {
                        "message": {
                          "type": "string",
                          "example": "Successfully uploaded the model card"
                        },
                        "model_card_id": {
                          "type": "string"
                        }
                      }
                    },
                    {
                      "type": "object",
                      "properties": {
                        "message": {
                          "type": "string",
                          "example": "Model card already exists"
                        },
                        "model_card_id": {
                          "type": "string"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        }
      }
    },
    "/modelcard/{mc_id}": {
      "get": {
        "summary": "Get model card",
        "description": "Retrieves a model card by its ID",
        "operationId": "get_model_card",
        "tags": ["Model Cards"],
        "parameters": [
          {
            "name": "mc_id",
            "in": "path",
            "description": "The unique identifier of the model card",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Model card successfully retrieved",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelCard"
                }
              }
            }
          },
          "400": {
            "description": "Model card not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "head": {
        "summary": "Get model card linkset",
        "description": "Returns an empty body with link information in the HTTP Link header",
        "operationId": "head_model_card",
        "tags": ["Model Cards"],
        "parameters": [
          {
            "name": "mc_id",
            "in": "path",
            "description": "The unique identifier of the model card",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success - Link headers contain the linkset information",
            "headers": {
              "Link": {
                "description": "Linkset relations for the model card",
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "404": {
            "description": "Model card not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      },
      "put": {
        "summary": "Update model card",
        "description": "Updates an existing model card in the Patra Knowledge Graph",
        "operationId": "put_update_model_card",
        "tags": ["Model Cards"],
        "parameters": [
          {
            "name": "mc_id",
            "in": "path",
            "description": "The unique identifier of the model card to update",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ModelCardUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Model card successfully updated or not found",
            "content": {
              "application/json": {
                "schema": {
                  "oneOf": [
                    {
                      "type": "object",
                      "properties": {
                        "message": {
                          "type": "string",
                          "example": "Successfully updated the model card"
                        },
                        "model_card_id": {
                          "type": "string"
                        }
                      }
                    },
                    {
                      "type": "object",
                      "properties": {
                        "message": {
                          "type": "string",
                          "example": "Model card not found"
                        },
                        "model_card_id": {
                          "type": "string"
                        }
                      }
                    }
                  ]
                }
              }
            }
          }
        }
      }
    },
...
    "/modelcard/{mc_id}/download_url": {
      "get": {
        "summary": "Get model download URL",
        "description": "Retrieves the download URL for a given model ID",
        "operationId": "get_model_download_url",
        "tags": ["Models"],
        "parameters": [
          {
            "name": "mc_id",
            "in": "path",
            "description": "The unique identifier of the model",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Model download URL successfully retrieved",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ModelLocation"
                }
              }
            }
          },
          "400": {
            "description": "Model not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
    "/modelcard/{mc_id}/deployments": {
      "get": {
        "summary": "Get model deployments",
        "description": "Retrieves all deployment information for a given model ID",
        "operationId": "get_model_deployments",
        "tags": ["Deployments"],
        "parameters": [
          {
            "name": "mc_id",
            "in": "path",
            "description": "The unique identifier of the model",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Model deployments successfully retrieved",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/Deployment"
                  }
                }
              }
            }
          },
          "400": {
            "description": "Deployments not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/modelcard/{mc_id}/location": {
      "put": {
        "summary": "Update model location",
        "description": "Updates the storage location information for a specific model",
        "operationId": "put_update_model_location",
        "tags": ["Models"],
        "parameters": [
          {
            "name": "mc_id",
            "in": "path",
            "description": "The unique identifier of the model",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ModelLocationUpdate"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Model location successfully updated",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "message": {
                      "type": "string",
                      "example": "Model location updated successfully"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request - Invalid location data or missing payload",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
    "/modelcard/id": {
      "post": {
        "summary": "Generate model ID",
        "description": "Generates a unique model ID for a given combination of author, name, and version",
        "operationId": "post_generate_model_id",
        "tags": ["Models"],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ModelIdRequest"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "New model ID successfully generated",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "pid": {
                      "type": "string",
                      "description": "Generated model ID"
                    }
                  }
                }
              }
            }
          },
          "400": {
            "description": "Bad request - Missing required parameters",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          },
          "409": {
            "description": "Conflict - Model ID already exists",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "pid": {
                      "type": "string",
                      "description": "Existing model ID"
                    }
                  }
                }
              }
            }
          },
          "500": {
            "description": "Internal server error - PID generation failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Error"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "ModelCard": {
        "type": "object",
        "description": "A model card containing information about a machine learning model",
        "additionalProperties": true
      },
      "ModelCardUpdate": {
        "type": "object",
        "description": "Data for updating an existing model card",
        "additionalProperties": true
      },
...
      "ModelLocation": {
        "type": "object",
        "description": "Information about where a model is stored",
        "additionalProperties": true
      },
      "ModelLocationUpdate": {
        "type": "object",
        "required": ["location"],
        "properties": {
          "location": {
            "type": "string",
            "format": "uri",
            "description": "The new storage location URL for the model"
          }
        }
      },
      "ModelIdRequest": {
        "type": "object",
        "required": ["author", "name", "version"],
        "properties": {
          "author": {
            "type": "string",
            "description": "The author/creator of the model",
            "example": "OpenAI"
          },
          "name": {
            "type": "string",
            "description": "The name of the model",
            "example": "GPT-4"
          },
          "version": {
            "type": "string",
            "description": "The version of the model",
            "example": "1.0"
          }
        }
      },
      "Deployment": {
        "type": "object",
        "description": "Information about a model deployment",
        "additionalProperties": true
      },
...
      "Error": {
        "type": "object",
        "required": ["error"],
        "properties": {
          "error": {
            "type": "string",
            "description": "Error message describing what went wrong"
          }
        }
      }
    }
  },
...
}